pyproject.toml
[project]
name = "tts-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "comtypes>=1.4; sys_platform == 'win32'",
    "loguru>=0.7",
    "pyyaml>=6.0",
]
